Refactoring (add base.models module)
